+++ lib/BackRest/Common/Lock.pm
@@ -325,7 +325,7 @@ push @EXPORT, qw(lockStop);
 ####################################################################################################################################
 # lockStopTest
 #
-# Test if a stop file exists for the current stanza or all stanza.
+# Test if a stop file exists for the current stanza or all stanzas.
 ####################################################################################################################################
 sub lockStopTest
 {
@@ -370,7 +370,7 @@ push @EXPORT, qw(lockStopTest);
 ####################################################################################################################################
 # lockStart
 #
-# Remove the stop file so processes can run
+# Remove the stop file so processes can run.
 ####################################################################################################################################
 sub lockStart
 {